Refactor UdpTestServer and move integration tests
-Move UdpTestServer to a utilities package and update its methods for thread safety. Refactor tests by separating integration tests into a dedicated class and simplifying unit tests for clarity and maintainability.
-Move Selector null check earlier to prevent making a network request when the response can't be processed.
